Crate kani

Re-exports

pub use arbitrary::Arbitrary;

pub use futures::block_on;

pub use futures::block_on_with_spawn;

pub use futures::spawn;

pub use futures::yield_now;

pub use futures::RoundRobin;

Modules

arbitrary

This module introduces the Arbitrary trait as well as implementation for primitive types and other std containers.

contracts

Kani implementation of function contracts.

futures

This module contains functions to work with futures (and async/.await) in Kani.

mem

This module contains functions useful for checking unsafe memory access.

slice

tuple

Support for arbitrary tuples where each element implements kani::Arbitrary. Tuples of size up to 12 are supported in this file.

vec

Macros

cover

A macro to check if a condition is satisfiable at a specific location in the code.

implies

implies!(premise => conclusion) means that if the premise is true, so must be the conclusion.

Functions

any

This creates an symbolic valid value of type T. You can assign the return value of this function to a variable that you want to make symbolic.

any_where

This creates a symbolic valid value of type T. The value is constrained to be a value accepted by the predicate passed to the filter. You can assign the return value of this function to a variable that you want to make symbolic.

assert

Creates an assertion of the specified condition and message.

assume

Creates an assumption that will be valid after this statement run. Note that the assumption will only be applied for paths that follow the assumption. If the assumption doesn’t hold, the program will exit successfully.

concrete_playback_run

NOP concrete_playback for type checking during verification mode.

cover

Creates a cover property with the specified condition and message.

Attribute Macros

ensures

Add a postcondition to this function.

modifies

Declaration of an explicit write-set for the annotated function.

proof

Marks a Kani proof harness

proof_for_contract

Designates this function as a harness to check a function contract.

requires

Add a precondition to this function.

should_panic

Specifies that a proof harness is expected to panic.**

solver

Select the SAT solver to use with CBMC for this harness

stub

Specify a function/method stub pair to use for proof harness

stub_verified

stub_verified(TARGET) is a harness attribute (to be used on proof or proof_for_contract function) that replaces all occurrences of TARGET reachable from this harness with a stub generated from the contract on TARGET.

unwind

Set Loop unwind limit for proof harnesses The attribute #[kani::unwind(arg)] can only be called alongside #[kani::proof]. arg - Takes in a integer value (u32) that represents the unwind value for the harness.

Derive Macros

Arbitrary

Allow users to auto generate Arbitrary implementations by using #[derive(Arbitrary)] macro.