Trait kani::arbitrary::Arbitrary

pub trait Arbitrary
where
    Self: Sized,
{
    // Required method
    fn any() -> Self;

    // Provided method
    fn any_array<const MAX_ARRAY_LENGTH: usize>() -> [Self; MAX_ARRAY_LENGTH]
       where [(); { _ }]: { ... }
}

This trait should be used to generate symbolic variables that represent any valid value of its type.

Required Methods

fn any() -> Self

Provided Methods

fn any_array<const MAX_ARRAY_LENGTH: usize>() -> [Self; MAX_ARRAY_LENGTH]
where [(); { _ }]:,

Object Safety

This trait is not object safe.