kani

Derive Macro BoundedArbitrary

#[derive(BoundedArbitrary)]
{
    // Attributes available to this derive:
    #[bounded]
}

Allow users to generate BoundedArbitrary implementations by using the #[derive(BoundedArbitrary)] macro.

Specifying bounded fields

By default, every field of a struct is required to derive Arbitrary, not BoundedArbitrary. This is because there is no general way to determine if a particular field should be bounded or not. You must annotate the fields you want to be bounded using the #[bounded] attribute.

For example, we only want the vector field in the following definition to be bounded.

#[derive(BoundedArbitrary)]
struct MyVector<T> {
    #[bounded]
    vector: Vec<T>,
    capacity: usize
}

Safety

Using BoundedArbitrary makes proofs incomplete. It is useful for increasing confidence that some code is correct, but does not prove that absolutely.