kani

Module float

This module contains functions useful for float-related checks

Functions

float_to_int_in_range

Returns whether the given float value satisfies the range condition of the to_int_unchecked methods, namely that the value after truncation is in range of the target Int