kani
Function check

pub const fn check(cond: bool, msg: &'static str)

👎Deprecated since 0.55.0: This API will be made private in future releases.

Creates a non-fatal property with the specified condition and message.

This check will not impact the program control flow even when it fails.

Example:

let x: bool = kani::any();
let y = !x;
kani::check(x || y, "ORing a boolean variable with its negation must be true");
kani::check(x == y, "A boolean variable is always different than its negation");
kani::cover!(true, "This should still be reachable");

Deprecated

This function was meant to be internal only, and it was added to Kani’s public interface by mistake. Thus, it will be made private in future releases. Instead, we recommend users to either use assert or cover to create properties they would like to verify.

See https://github.com/model-checking/kani/issues/3561 for more details.