kani
Module invariant

This module introduces the Invariant trait as well as its implementation for primitive types.

Traits

Invariant

This trait should be used to specify and check type safety invariants for a type. For type invariants, we refer to the definitions in the Rust’s Unsafe Code Guidelines Reference: https://rust-lang.github.io/unsafe-code-guidelines/glossary.html#validity-and-safety-invariant