Trait kani::arbitrary::Arbitrary

pub trait Arbitrary {
    fn any() -> Self;
}

This trait should be used to generate symbolic variables that represent any valid value of its type.

Required Methods

fn any() -> Self