kani::futures

Function spawn

pub fn spawn<F: Future<Output = ()> + Sync + 'static>(fut: F) -> JoinHandle 

Spawns a task on the current global executor (which is set by block_on_with_spawn)

This function can only be called inside a future passed to block_on_with_spawn.